Separating basic siren functions into their own behaviour.
