+++ components/d2l-quick-eval/behaviors/d2l-quick-eval-siren-helper-behavior.js
@@ -1,6 +1,5 @@
 import {Rels, Classes} from 'd2l-hypermedia-constants';
-import 'd2l-polymer-siren-behaviors/store/entity-behavior.js';
-import 'd2l-polymer-siren-behaviors/store/siren-action-behavior.js';
+import './d2l-siren-helper-behavior.js';
 
 window.D2L = window.D2L || {};
 window.D2L.PolymerBehaviors = window.D2L.PolymerBehaviors || {};
@@ -11,29 +10,6 @@ window.D2L.PolymerBehaviors.QuickEval = window.D2L.PolymerBehaviors.QuickEval ||
 * @polymerBehavior
 */
 D2L.PolymerBehaviors.QuickEval.D2LQuickEvalSirenHelperBehaviorImpl = {
-
-	_myEntityStoreFetch : function(url) {
-		return window.D2L.Siren.EntityStore.fetch(url, this.token);
-	},
-
-	_followLink: function(entity, rel) {
-		const href = this._getHref(entity, rel);
-		return this._followHref(href);
-	},
-
-	_getHref: function(entity, rel) {
-		if (entity && entity.hasLinkByRel && entity.hasLinkByRel(rel)) {
-			return entity.getLinkByRel(rel).href;
-		}
-		return '';
-	},
-
-	_followHref: function(href) {
-		if (href) {
-			return this._myEntityStoreFetch(href);
-		}
-		return Promise.resolve();
-	},
 
 	_getMasterTeacherPromise: function(entity, item) {
 		return this._followLink(entity, Rels.organization)
@@ -156,79 +132,10 @@ D2L.PolymerBehaviors.QuickEval.D2LQuickEvalSirenHelperBehaviorImpl = {
 		}
 		return '';
 	},
-
-	_performSirenActionWithQueryParams: function(action, customParams) {
-		const url = new URL(action.href, window.location.origin);
-
-		if (!action.fields) {
-			action.fields = [];
-		}
-
-		url.searchParams.forEach(function(value, key) {
-			if (!action.fields.filter(x => x.name === key)[0]) {
-				action.fields.push({name: key, value: value, type: 'hidden'});
-			}
-		});
-
-		if (customParams) {
-			Object.keys(customParams).forEach(function(paramName) {
-				action.fields.push({name: paramName, value: customParams[paramName], type: 'hidden'});
-			});
-		}
-
-		return this.performSirenAction(action);
-	},
-
-	_getExtraParams: function(url) {
-		if (!url || url === '') return [];
-
-		const extraParams = [];
-
-		const filterVal = this._getQueryStringParam('filter', url);
-		if (filterVal) {
-			extraParams.push(
-				{
-					name: 'filter',
-					value: filterVal
-				}
-			);
-		}
-		const sortVal = this._getQueryStringParam('sort', url);
-		if (sortVal) {
-			extraParams.push(
-				{
-					name: 'sort',
-					value: sortVal
-				}
-			);
-		}
-
-		return extraParams;
-	},
-
-	_getQueryStringParam: function(name, url) {
-		const parsedUrl = new window.URL(url);
-		return parsedUrl.searchParams.get(name);
-	},
-
-	_buildRelativeUri: function(url, extraParams) {
-		if (extraParams.length === 0) {
-			return url;
-		}
-
-		const parsedUrl = new window.URL(url, 'https://notused.com');
-
-		extraParams.forEach(param => {
-			parsedUrl.searchParams.set(param.name, param.value);
-		});
-
-		return parsedUrl.pathname + parsedUrl.search;
-	}
 };
 
 /** @polymerBehavior */
 D2L.PolymerBehaviors.QuickEval.D2LQuickEvalSirenHelperBehavior = [
-	D2L.PolymerBehaviors.Siren.EntityBehavior,
-	D2L.PolymerBehaviors.Siren.SirenActionBehavior,
+	D2L.PolymerBehaviors.Siren.D2LSirenHelperBehavior,
 	D2L.PolymerBehaviors.QuickEval.D2LQuickEvalSirenHelperBehaviorImpl
 ];
