+++ test/d2l-quick-eval/behaviors/d2l-quick-eval-siren-helper-behavior.html
@@ -5,13 +5,13 @@
 		<meta name="viewport" content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes">
 
 		<title>d2l-quick-eval test</title>
-		<script src="../../../@webcomponents/webcomponentsjs/webcomponents-bundle.js"></script>
-		<script src="../../../wct-browser-legacy/browser.js"></script>
+		<script src="../../../../@webcomponents/webcomponentsjs/webcomponents-bundle.js"></script>
+		<script src="../../../../wct-browser-legacy/browser.js"></script>
 
 		<!-- For IE11 -->
-		<script src="../../../lie/dist/lie.polyfill.min.js"></script>
-		<script type="module" src="../../../whatwg-fetch/fetch.js"></script>
-		<script type="module" src="../../../url-polyfill/url-polyfill.js"></script>
+		<script src="../../../../lie/dist/lie.polyfill.min.js"></script>
+		<script type="module" src="../../../../whatwg-fetch/fetch.js"></script>
+		<script type="module" src="../../../../url-polyfill/url-polyfill.js"></script>
 		<script type="module" src="./d2l-quick-eval-siren-helper-behavior-test-component.js"></script>
 	</head>
 	<body>
