+++ test/index.html
@@ -20,8 +20,10 @@
 			'd2l-quick-eval/d2l-quick-eval.html?dom=shadow',
 			'd2l-quick-eval/d2l-quick-eval-activities-list.html?wc-shadydom=true&wc-ce=true',
 			'd2l-quick-eval/d2l-quick-eval-activities-list.html?dom=shadow',
-			'd2l-quick-eval/d2l-quick-eval-siren-helper-behavior.html?wc-shadydom=true&wc-ce=true',
-			'd2l-quick-eval/d2l-quick-eval-siren-helper-behavior.html?dom=shadow',
+			'd2l-quick-eval/behaviors/d2l-quick-eval-siren-helper-behavior.html?wc-shadydom=true&wc-ce=true',
+			'd2l-quick-eval/behaviors/d2l-quick-eval-siren-helper-behavior.html?dom=shadow',
+			'd2l-quick-eval/behaviors/d2l-siren-helper-behavior.html?wc-shadydom=true&wc-ce=true',
+			'd2l-quick-eval/behaviors/d2l-siren-helper-behavior.html?dom=shadow',
 			'd2l-activity-evaluation-icon/d2l-activity-evaluation-icon-base.html?wc-shadydom=true&wc-ce=true',
 			'd2l-activity-evaluation-icon/d2l-activity-evaluation-icon-base.html?dom=shadow',
 			'd2l-quick-eval/d2l-quick-eval-activities-list-sorting.html?wc-shadydom=true&wc-ce=true',
